Created Package package
Created Package package to reduce code duplication;
